updated dbyll version.
the old version was a mess, full of unfinished bad workarounds
for the rigid classless template in a theme that expected a
different section order.

this new version uses CSS Grid to fix everything in a very clean
manner. with that we also get easy and solid responsiveness.

there are also improvements in link colors and other minor stuff.